+++ src/FileData/Curiosity.FileDataReaderWriters/Curiosity.FileDataReaderWriters.csproj
