+++ src/FileData/Curiosity.FileDataReaderWriters/Readers/IFileDataReader.cs
@@ -1,4 +1,4 @@
-﻿namespace Curiosity.Utils.FileDataReaderWriters.Readers;
+﻿namespace Curiosity.FileDataReaderWriters.Readers;
 
 /// <summary>
 /// Общий интерфейс для классов, которые читает загруженные файлы построчно.
