+++ src/FileData/Curiosity.FileDataReaderWriters/Resources/Strings.en.resx
