+++ src/FileData/Curiosity.FileDataReaderWriters/Resources/Strings.resx
