+++ src/FileData/Curiosity.FileDataReaderWriters/Style/FormatSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Curiosity.Utils.FileDataReaderWriters.Style;
+﻿namespace Curiosity.FileDataReaderWriters.Style;
 
 /// <summary>
 /// Настройки формата ячеек
