+++ src/FileData/Curiosity.FileDataReaderWriters/Writers/NpoiXlsMultiFileWriter.cs
@@ -1,37 +1,50 @@
-﻿using Curiosity.Utils.FileDataReaderWriters.Style;
+﻿using Curiosity.FileDataReaderWriters.Helpers;
+using Curiosity.FileDataReaderWriters.Style;
 using ICSharpCode.SharpZipLib.Zip;
 using Microsoft.Extensions.Logging;
 using NPOI.SS.UserModel;
 using NPOI.XSSF.Streaming;
 
-namespace Curiosity.Utils.FileDataReaderWriters.Writers;
+namespace Curiosity.FileDataReaderWriters.Writers;
 
 /// <summary>
-/// Класс для записи данных в xlsx / xls формате для NPOI
+/// Класс для записи данных в xlsx / xls формате для NPOI с разбитием на несколько файлов
 /// </summary>
-public class NpoiXlsFileWriter : IFileWriter
+public class NpoiXlsMultiFileWriter : IFileWriter
 {
+    private readonly ILogger _logger;
+    private          IReadOnlyList<CellData>? _headers;
     private          int _currentRow = 0;
     private          int _currentCell = 0;
-    private readonly string _outputFilePath;
-    private readonly ILogger _logger;
+    private          int _partNumber = 1;
 
     private readonly SXSSFWorkbook _workbook;
-    private readonly ISheet _sheet;
     private readonly ICellStyle _dataStyle;
+    private          ISheet _sheet;
     private          FileStream _fileStream;
     
     private          Dictionary<int, ICellStyle> _cellStyles = new();
+    
+    private readonly string _savePath;
+    private readonly string _fileName;
 
     /// <summary>
-    /// Конструктор для записи эксель файла через NPOI
+    /// Конструктор для записи множества эксель файлов через NPOI
     /// </summary>
-    /// <param name="outputFilePath">Адрес итогового файла</param>
+    /// <param name="logger"></param>
+    /// <param name="fileName">Название файла</param>
+    /// <param name="savePath">Директория сохранения</param>
     /// <param name="rowAccessWindowSize">Количество строк, находящихся в памяти писателя</param>
-    public NpoiXlsFileWriter(string outputFilePath, ILogger logger, int rowAccessWindowSize = 100)
+    public NpoiXlsMultiFileWriter(string savePath, string fileName, ILogger logger, int rowAccessWindowSize = 100)
     {
-        _outputFilePath = outputFilePath;
+        if (String.IsNullOrWhiteSpace(savePath))
+            throw new ArgumentException("Value cannot be null or whitespace.", nameof(savePath));
+        if (String.IsNullOrWhiteSpace(fileName))
+            throw new ArgumentException("Value cannot be null or whitespace.", nameof(fileName));
+
+        _savePath = savePath;
         _logger = logger;
+        _fileName = fileName;
         _workbook = new SXSSFWorkbook(rowAccessWindowSize);
         _workbook.UseZip64 = UseZip64.On;
         _sheet = _workbook.CreateSheet();
@@ -41,7 +54,7 @@ public class NpoiXlsFileWriter : IFileWriter
         _dataStyle = _workbook.CreateCellStyle();
         _dataStyle.DataFormat = format.GetFormat("yyyy-MM-dd HH:mm:ss");
     }
-
+    
     public int AddFormat(FormatSettings formatSettings)
     {
         var font = _workbook.CreateFont();
@@ -79,18 +92,17 @@ public class NpoiXlsFileWriter : IFileWriter
 
         return formatNumber;
     }
-
+    
     public void AddHeaders(IReadOnlyList<CellData> data)
     {
-        if (data is null)
-            throw new ArgumentNullException(nameof(data));
+        _headers = data ?? throw new ArgumentNullException(nameof(data));
 
         for (var i = 0; i < data.Count; i++)
         {
             var datum = data[i];
             Append(datum.Value, datum.Format);
         }
-
+        
         EndLine();
     }
 
@@ -162,11 +174,11 @@ public class NpoiXlsFileWriter : IFileWriter
                 _logger.LogWarning("Используем обычный ToString() (Type={Type}, Value={Value})", value.GetType(), value);
                 cell.SetCellValue(value.ToString());
                 break;
-        } 
+        }
         
         if (format is not null)
             cell.CellStyle = _cellStyles[(int)format];
-
+        
         _currentCell++;
     }
 
@@ -174,19 +186,50 @@ public class NpoiXlsFileWriter : IFileWriter
     {
         _currentRow++;
         _currentCell = 0;
+
+        // если лимит - сбросим данные в файл
+        if (_currentRow >= ExcelConstants.XlsxRowsMax)
+            Flush();
     }
 
     public void Flush()
     {
-    }
+        // если ничего нет
+        // или только хедеры и это 2ой файл - выходим
+        // (первый пустой файл с хедерами - сохраняем)
+        if (_currentRow == 0 || 
+            (_headers != null && _currentRow == 1 && _partNumber > 1))
+            return;
 
-    /// <summary>
-    /// Записываем данные в файл и закрываем стрим.
-    /// </summary>
-    public void Dispose()
-    {
-        _fileStream = File.Open(_outputFilePath, FileMode.OpenOrCreate, FileAccess.ReadWrite);
+        // сгенерим имя
+        var fileName = _fileName;
+        if (_partNumber > 1)
+        {
+            fileName =
+                $"{Path.GetFileNameWithoutExtension(_fileName)}_part_{_partNumber}{Path.GetExtension(_fileName)}";
+        }
+
+        var outputFilePath = Path.Combine(_savePath, fileName);
+
+        // сохраним
+        _logger.LogDebug($"Сохраняем файл \"{outputFilePath}\"...");
+        _fileStream = File.Open(outputFilePath, FileMode.OpenOrCreate, FileAccess.Write);
         _workbook.Write(_fileStream);
         _fileStream.Close();
+        _logger.LogInformation($"Файл \"{outputFilePath}\" успешно сохранён");
+
+        // обнулим состояние
+        _workbook.RemoveSheetAt(0);
+        _sheet = _workbook.CreateSheet();
+        _partNumber++;
+        _currentRow = 0;
+        if (_headers != null)
+            AddHeaders(_headers);
+    }
+
+    public void Dispose()
+    {
+        if (_fileStream is not null)
+            Flush();
     }
 }
